Modify event item component
